Refactor produce editing to use modal
Replaces the separate edit page and form component with an EditProduceModal, allowing produce items to be edited directly from the ProduceItem table via a modal dialog. Updates ProduceItem to use the modal and adds custom button styles in buttons.css for improved UI consistency.
