--- src/app/edit/[id]/page.tsx
@@ -1,33 +1,0 @@
-import { getServerSession } from 'next-auth';
-import { notFound } from 'next/navigation';
-import { Produce } from '@prisma/client';
-import authOptions from '@/lib/authOptions';
-import { loggedInProtectedPage } from '@/lib/page-protection';
-import { prisma } from '@/lib/prisma';
-import EditProduceForm from '@/components/EditProduceForm';
-
-export default async function EditProducePage({ params }: { params: { id: string | string[] } }) {
-  // Protect the page, only logged in users can access it.
-  const session = await getServerSession(authOptions);
-  loggedInProtectedPage(
-    session as {
-      user: { email: string; id: string; randomKey: string };
-      // eslint-disable-next-line @typescript-eslint/comma-dangle
-    } | null,
-  );
-  const id = Number(Array.isArray(params?.id) ? params?.id[0] : params?.id);
-  // console.log(id);
-  const produce: Produce | null = await prisma.produce.findUnique({
-    where: { id },
-  });
-  // console.log(stuff);
-  if (!produce) {
-    return notFound();
-  }
-
-  return (
-    <main>
-      <EditProduceForm produce={produce} />
-    </main>
-  );
-}
