+++ src/components/ProduceItem.tsx
@@ -1,20 +1,38 @@
 import { Produce } from '@prisma/client';
-import Link from 'next/link';
+import { useState } from 'react';
+import { Button } from 'react-bootstrap';
+import EditProduceModal from './EditProduceModal';
+import '../styles/buttons.css';
 
-const ProduceItem = ({ id, name, quantity, unit, type, location, expiration }: Produce) => (
-  <tr>
-    <td>{name}</td>
-    <td>{type}</td>
-    <td>{location}</td>
-    <td>
-      {quantity.toString()}
-      {unit ? ` ${unit}` : ''}
-    </td>
-    <td>{expiration ? new Date(expiration).toLocaleDateString() : 'N/A'}</td>
-    <td>
-      <Link href={`/edit/${id}`}>Edit</Link>
-    </td>
-  </tr>
-);
+const ProduceItem = ({ id, name, quantity, unit, type, location, expiration, owner, image }: Produce) => {
+  const [showModal, setShowModal] = useState(false);
+
+  return (
+    <>
+      <tr>
+        <td>{name}</td>
+        <td>{type}</td>
+        <td>{location}</td>
+        <td>
+          {quantity.toString()}
+          {unit ? ` ${unit}` : ''}
+        </td>
+        <td>{expiration ? new Date(expiration).toLocaleDateString() : 'N/A'}</td>
+        <td>
+          <Button className="btn-edit" onClick={() => setShowModal(true)}>
+            Edit
+          </Button>
+        </td>
+      </tr>
+
+      {/* Modal component for editing produce item */}
+      <EditProduceModal
+        show={showModal}
+        onHide={() => setShowModal(false)}
+        produce={{ id, name, quantity, unit, type, location, expiration, owner, image }}
+      />
+    </>
+  );
+};
 
 export default ProduceItem;
