+++ src/styles/buttons.css
@@ -1,0 +1,54 @@
+.btn-edit {
+    width: 100%;
+    padding: 12px;
+    background-color: var(--hunter-green);
+    color: white;
+    font-size: 1rem;
+    font-weight: bold;
+    border: none;
+    border-radius: 8px;
+    transition: background-color 0.3s ease, transform 0.2s ease;
+    cursor: pointer;
+    text-transform: uppercase;
+}
+
+.btn-edit:hover {
+    background-color: var(--fern-green);
+    transform: translateY(-2px);
+}
+
+.btn-submit {
+    width: 100%;
+    padding: 12px;
+    background-color: var(--brunswick-green);
+    color: white;
+    font-size: 1rem;
+    font-weight: bold;
+    border: none;
+    border-radius: 8px;
+    transition: background-color 0.3s ease, transform 0.2s ease;
+    cursor: pointer;
+    text-transform: uppercase;
+}
+
+.btn-submit:hover {
+    background-color: var(--fern-green);
+    transform: translateY(-2px);
+}
+
+.btn-reset {
+    width: 100%;
+    padding: 12px;
+    font-size: 1rem;
+    font-weight: bold;
+    border: none;
+    border-radius: 8px;
+    transition: background-color 0.3s ease, transform 0.2s ease;
+    cursor: pointer;
+    text-transform: uppercase;
+}
+
+.btn-reset:hover {
+    background-color: red;
+    transform: translateY(-2px);
+}
